requirements.txt
requests~=2.26.0
PyYAML~=6.0
pandas~=1.3.4
numpy~=1.21.3